test: remove AutoFixture and NSubstitute infrastructure
