+++ Directory.Packages.props
@@ -78,12 +78,8 @@
     <PackageVersion Include="Verify.XunitV3" Version="31.28.0" />
   </ItemGroup>
   <ItemGroup Label="Testing">
-    <PackageVersion Include="AutoFixture" Version="4.18.1" />
-    <PackageVersion Include="AutoFixture.AutoNSubstitute" Version="4.18.1" />
-    <PackageVersion Include="AutoFixture.Xunit3" Version="4.19.0" />
     <PackageVersion Include="AwesomeAssertions" Version="9.5.0" />
     <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="18.8.1" />
-    <PackageVersion Include="NSubstitute" Version="6.2.0" />
     <PackageVersion Include="xunit.runner.visualstudio" Version="3.1.5" />
     <PackageVersion Include="xunit.v3.mtp-v2" Version="3.2.2" />
   </ItemGroup>
@@ -97,7 +93,6 @@
     <PackageVersion Include="Compono.XunitV3" Version="0.8.0" />
     <PackageVersion Include="Compono.Http" Version="0.8.0" />
     <PackageVersion Include="Compono.TestDoubles" Version="0.8.0" />
-    <PackageVersion Include="Compono.NSubstitute" Version="0.8.0" />
     <PackageVersion Include="Compono.Bogus" Version="0.8.0" />
   </ItemGroup>
   <ItemGroup Label="Reference Assemblies">
