+++ test/AlexaVoxCraft.InSkillPurchasing.Tests/AlexaVoxCraft.InSkillPurchasing.Tests.csproj
@@ -32,13 +32,6 @@
     </ItemGroup>
 
     <ItemGroup>
-      <Using Remove="AutoFixture" />
-      <Using Remove="AutoFixture.Xunit3" />
-      <Using Remove="NSubstitute" />
-      <Using Remove="AlexaVoxCraft.TestKit" />
-      <Using Remove="AlexaVoxCraft.TestKit.Attributes" />
-    </ItemGroup>
-    <ItemGroup>
         <!-- Embed all JSON fixtures under Examples/ -->
         <EmbeddedResource Include="Examples\**\*.json"/>
     </ItemGroup>
