+++ test/AlexaVoxCraft.Model.Apl.Legacy.Tests/AlexaVoxCraft.Model.Apl.Legacy.Tests.csproj
@@ -20,12 +20,7 @@
     </ItemGroup>
 
     <ItemGroup>
-        <Using Remove="AutoFixture" />
-        <Using Remove="AutoFixture.Xunit3" />
-        <Using Remove="NSubstitute" />
         <Using Remove="AwesomeAssertions" />
-        <Using Remove="AlexaVoxCraft.TestKit" />
-        <Using Remove="AlexaVoxCraft.TestKit.Attributes" />
     </ItemGroup>
 
     <ItemGroup>
