+++ test/AlexaVoxCraft.Model.Legacy.Tests/AlexaVoxCraft.Model.Legacy.Tests.csproj
@@ -27,13 +27,6 @@
     </ItemGroup>
 
     <ItemGroup>
-        <Using Remove="AutoFixture" />
-        <Using Remove="AutoFixture.Xunit3" />
-        <Using Remove="NSubstitute" />
-        <Using Remove="AlexaVoxCraft.TestKit" />
-        <Using Remove="AlexaVoxCraft.TestKit.Attributes" />
-    </ItemGroup>
-    <ItemGroup>
         <None Remove="Examples\AskForPermissionConsent.json"/>
         <None Update="Examples\TemplateImageBasic.json">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
