--- test/AlexaVoxCraft.TestKit/AlexaVoxCraft.TestKit.csproj
@@ -1,29 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <LangVersion>default</LangVersion>
-        <IsPackable>false</IsPackable>
-        <TestingPlatformDotnetTestSupport>true</TestingPlatformDotnetTestSupport>
-        <UseMicrosoftTestingPlatformRunner>true</UseMicrosoftTestingPlatformRunner>
-        <IsTestProject>false</IsTestProject>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="AutoFixture" />
-      <PackageReference Include="AutoFixture.AutoNSubstitute" />
-      <PackageReference Include="AutoFixture.Xunit3" />
-      <PackageReference Include="AwesomeAssertions" />
-        <PackageReference Include="NSubstitute">
-            <NoWarn>NU1608</NoWarn>
-        </PackageReference>
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\..\src\AlexaVoxCraft.MediatR.Lambda\AlexaVoxCraft.MediatR.Lambda.csproj" />
-      <ProjectReference Include="..\..\src\AlexaVoxCraft.MediatR\AlexaVoxCraft.MediatR.csproj" />
-      <ProjectReference Include="..\..\src\AlexaVoxCraft.Model\AlexaVoxCraft.Model.csproj" />
-    </ItemGroup>
-
-</Project>
