--- test/AlexaVoxCraft.TestKit/Attributes/AlexaVoxCraftAutoDataAttribute.cs
@@ -1,16 +1,0 @@
-namespace AlexaVoxCraft.TestKit.Attributes;
-
-public class AlexaVoxCraftAutoDataAttribute() : AutoDataAttribute(CreateFixture)
-{
-    internal static IFixture CreateFixture()
-    {
-        return BaseFixtureFactory.CreateFixture(fixture =>
-        {
-        });
-    }
-}
-
-public class InlineAlexaVoxCraftAutoDataAttribute(params object[] values)
-    : InlineAutoDataAttribute(AlexaVoxCraftAutoDataAttribute.CreateFixture, values)
-{
-}
