--- test/AlexaVoxCraft.TestKit/RequestSpecifications/LambdaContextSpecification.cs
@@ -1,7 +1,0 @@
-using Amazon.Lambda.Core;
-
-namespace AlexaVoxCraft.TestKit.RequestSpecifications;
-
-public class LambdaContextSpecification : BaseTypedSpecification<ILambdaContext>
-{
-}
