--- test/AlexaVoxCraft.TestKit/RequestSpecifications/LogEventSpecification.cs
@@ -1,7 +1,0 @@
-using Serilog.Events;
-
-namespace AlexaVoxCraft.TestKit.RequestSpecifications;
-
-public class LogEventSpecification : BaseTypedSpecification<LogEvent>
-{
-}
