--- test/AlexaVoxCraft.TestKit/RequestSpecifications/OptionsSpecification.cs
@@ -1,8 +1,0 @@
-using AlexaVoxCraft.MediatR.DI;
-using Microsoft.Extensions.Options;
-
-namespace AlexaVoxCraft.TestKit.RequestSpecifications;
-
-public class OptionsSpecification : BaseTypedSpecification<IOptions<SkillServiceConfiguration>>
-{
-}
