--- test/AlexaVoxCraft.TestKit/RequestSpecifications/RequestHandlerDelegateSpecification.cs
@@ -1,7 +1,0 @@
-using AlexaVoxCraft.MediatR.Pipeline;
-
-namespace AlexaVoxCraft.TestKit.RequestSpecifications;
-
-public class RequestHandlerDelegateSpecification : BaseTypedSpecification<RequestHandlerDelegate>
-{
-}
