--- test/AlexaVoxCraft.TestKit/RequestSpecifications/ServiceCollectionSpecification.cs
@@ -1,7 +1,0 @@
-using Microsoft.Extensions.DependencyInjection;
-
-namespace AlexaVoxCraft.TestKit.RequestSpecifications;
-
-public class ServiceCollectionSpecification : BaseTypedSpecification<IServiceCollection>
-{
-}
