--- test/AlexaVoxCraft.TestKit/RequestSpecifications/ServiceProviderSpecification.cs
@@ -1,5 +1,0 @@
-namespace AlexaVoxCraft.TestKit.RequestSpecifications;
-
-public class ServiceProviderSpecification : BaseTypedSpecification<IServiceProvider>
-{
-}
