--- test/AlexaVoxCraft.TestKit/RequestSpecifications/SkillRequestFactorySpecification.cs
@@ -1,7 +1,0 @@
-using AlexaVoxCraft.MediatR;
-
-namespace AlexaVoxCraft.TestKit.RequestSpecifications;
-
-public class SkillRequestFactorySpecification : BaseTypedSpecification<SkillRequestFactory>
-{
-}
