--- test/AlexaVoxCraft.TestKit/RequestSpecifications/SkillServiceConfigurationSpecification.cs
@@ -1,7 +1,0 @@
-using AlexaVoxCraft.MediatR.DI;
-
-namespace AlexaVoxCraft.TestKit.RequestSpecifications;
-
-public class SkillServiceConfigurationSpecification : BaseTypedSpecification<SkillServiceConfiguration>
-{
-}
