+++ test/Directory.Build.props
@@ -4,12 +4,7 @@
     </PropertyGroup>
     <ItemGroup>
         <Using Include="Xunit"/>
-        <Using Include="AutoFixture"/>
-        <Using Include="AutoFixture.Xunit3"/>
-        <Using Include="NSubstitute"/>
         <Using Include="AwesomeAssertions"/>
-        <Using Include="AlexaVoxCraft.TestKit" />
-        <Using Include="AlexaVoxCraft.TestKit.Attributes" />
     </ItemGroup>
     <ItemGroup Condition="'$(IsTestProject)' == 'true'">
         <PackageReference Include="Microsoft.NET.Test.Sdk" />
